fix(plugin-openwhisk): separate out the editor parts from plugin-openwhisk Fixes #437
also fixes missing red asterisk after `new`
Fixes #441
