+++ plugins/plugin-openwhisk-editor-extensions/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@kui-shell/plugin-openwhisk-editor-extensions",
+  "version": "0.0.1",
+  "description": "Kui plugin for OpenWhisk that extends plugin-editor",
+  "license": "Apache-2.0",
+  "author": "Nick Mitchell",
+  "homepage": "https://github.com/IBM/kui#readme",
+  "bugs": {
+    "url": "https://github.com/IBM/kui/issues/new"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/IBM/kui.git"
+  },
+  "keywords": [
+    "kui",
+    "plugin",
+    "openwhisk"
+  ],
+  "main": "plugin.js",
+  "scripts": {
+    "prepack": "rm -rf dist && mkdir dist && tar -C ../../build/plugins/$(basename `pwd`)/src --exclude node_modules -cf - . | tar -C dist -xf - && find dist -type l -exec rm {} \\; && cp -a dist/* .",
+    "postpack": "for i in dist/*; do if [ -d $i ]; then for j in $i/*; do rm -rf ./`basename $i`/`basename $j`; done; else rm -rf ./`basename $i`; fi; done && rm -rf dist"
+  },
+  "dependencies": {},
+  "kui": {
+    "headless": false
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
