+++ plugins/plugin-openwhisk-editor-extensions/src/lib/placeholders.ts
