+++ plugins/plugin-openwhisk-editor-extensions/src/plugin.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2018 IBM Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as Debug from 'debug'
+const debug = Debug('plugins/openwhisk-editor-extensions/loader')
+
+import newCmd from './lib/cmds/new'
+import composeCmd from './lib/cmds/compose'
+
+export default async (commandTree, prequire) => {
+  const wsk = await prequire('plugin-openwhisk')
+
+  await newCmd(commandTree, wsk)
+  await composeCmd(commandTree, prequire)
+}
